初步拆分文件，实现队列功能 待做： 1. 解决编译报错 2. 队列资源竞争问题

+++ User/my_rtos/Inc/queue.h
@@ -1,13 +1,12 @@
 #pragma once
-
-#include "sparrow.h"
+#include "config.h"
 
 Class (Queue_t) {
     uint8_t *startPoint;
     uint8_t *endPoint;
     uint8_t *readPoint;
     uint8_t *writePoint;
-    uint8_t messageNumber;
+    uint32_t messageNumber;
     uint32_t sendTable;
     uint32_t receiveTable;
     uint32_t nodeSize;
